Create driver_station.py - to send serial commands
This file is a python script to be run on the driver's station laptop.

It is used something like this:  python driver_station.py COM10   <enter>

When you launch it, it'll mess around and print some status and tell you can do input, but it will ask you to count to 10. If you don't wait a bit before you start, it will hang, so just give it 10 seconds after it says it's all ok.

To exit the program, hit the "q" key.  Whatever keys you hit will be dispatched into your program on the Arduino if you used the framework provided to the team for the Fall of 2014 training.

This file requires that you have python 2.7 or later installed with the pyserial library and the msvcrt library (the latter should be automatic but the serial one requires that you install it.