Create provisional_template_for_arduino_agent_code.template
This program is not meant to be run as anything more than a test of the driver_station script.  To do that create a project and copy this code into it. The expected behavior is that you can type characters into the driver_station.py and have them echoed twice followed by an "ok" at the driver_station.py terminal.

However, this will be the base of the eventual framework for the arduino code in the Fall of 2014 training program at PyroTech.